Pivoting bot-type, and making things work.
A bot that modifies the UI is tricky, because you have to pass back the
text and also the action, which complicates the skeleton code and makes
it harder to extend. Replace it with a simpler bot that can do
modifications to words for you.

Also, modify our docker-run-server.sh script to restart the server when
a file changes, via nodemon.